A bunch more static stuff moved to sound package.
