+++ src/main/java/org/runejs/client/Class40_Sub3.java
@@ -30,7 +30,6 @@ import org.runejs.client.sound.StaticAudio;
 public class Class40_Sub3 extends Node {
     public static ImageRGB[] aClass40_Sub5_Sub14_Sub4Array2019;
     public static int ticksPerLoop;
-    public static int[] soundDelay = new int[50];
     public static int anInt2024 = 1;
     public static boolean showIconsRedrawnText = false;
     public static int loginStatus = 0;
@@ -49,15 +48,15 @@ public class Class40_Sub3 extends Node {
 
 
     public static void startup(int arg0) {
-        if (Class40_Sub5_Sub6.loadingPercent == 0) {
+        if (MovedStatics.loadingPercent == 0) {
             Npc.currentScene = new Scene(MovedStatics.tile_height);
             for (int i = 0; i < 4; i++)
                 Landscape.currentCollisionMap[i] = new CollisionMap(104, 104);
             MovedStatics.minimapImage = new ImageRGB(512, 512);
             MovedStatics.anInt1607 = 5;
-            Class40_Sub5_Sub6.loadingPercent = 20;
+            MovedStatics.loadingPercent = 20;
             Native.currentLoadingText = English.startingGameEngine;
-        } else if (Class40_Sub5_Sub6.loadingPercent == 20) {
+        } else if (MovedStatics.loadingPercent == 20) {
 //            int[] is = new int[9];
 //            for (int i = 0; i < 9; i++) {
 //                int i_0_ = 15 + 32 * i + 128;
@@ -69,8 +68,8 @@ public class Class40_Sub3 extends Node {
             ScreenController.setBounds();
             MovedStatics.anInt1607 = 10;
             Native.currentLoadingText = English.preparedVisibilityMap;
-            Class40_Sub5_Sub6.loadingPercent = 30;
-        } else if (Class40_Sub5_Sub6.loadingPercent == 30) {
+            MovedStatics.loadingPercent = 30;
+        } else if (MovedStatics.loadingPercent == 30) {
             CacheArchive.skeletonCacheArchive = CacheArchive.loadArchive(0, true, false, true);
             CacheArchive.skinDefinitionCacheArchive = CacheArchive.loadArchive(1, true, false, true);
             CacheArchive.gameDefinitionsCacheArchive = CacheArchive.loadArchive(2, true, true, false);
@@ -84,10 +83,10 @@ public class Class40_Sub3 extends Node {
             CacheArchive.huffmanCacheArchive = CacheArchive.loadArchive(10, true, false, true);
             CacheArchive.jingleCacheArchive = CacheArchive.loadArchive(11, true, false, true);
             CacheArchive.clientScriptCacheArchive = CacheArchive.loadArchive(12, true, false, true);
-            Class40_Sub5_Sub6.loadingPercent = 40;
+            MovedStatics.loadingPercent = 40;
             Native.currentLoadingText = English.connectingToUpdateServer;
             MovedStatics.anInt1607 = 20;
-        } else if (Class40_Sub5_Sub6.loadingPercent == 40) {
+        } else if (MovedStatics.loadingPercent == 40) {
             int i = 0;
             i += CacheArchive.skeletonCacheArchive.getPercentLoaded() * 5 / 100;
             i += 5 * CacheArchive.skinDefinitionCacheArchive.getPercentLoaded() / 100;
@@ -103,7 +102,7 @@ public class Class40_Sub3 extends Node {
             i += 5 * CacheArchive.jingleCacheArchive.getPercentLoaded() / 100;
             i += 5 * CacheArchive.clientScriptCacheArchive.getPercentLoaded() / 100;
             if (i == 100) {
-                Class40_Sub5_Sub6.loadingPercent = 45;
+                MovedStatics.loadingPercent = 45;
                 Native.currentLoadingText = English.loadedUpdateList;
                 MovedStatics.anInt1607 = 30;
             } else {
@@ -111,14 +110,14 @@ public class Class40_Sub3 extends Node {
                     Native.currentLoadingText = English.checkingForUpdates + i + Native.percent;
                 MovedStatics.anInt1607 = 30;
             }
-        } else if (Class40_Sub5_Sub6.loadingPercent == 45) {
-            MovedStatics.method446(Main.signlink, 0, !VertexNormal.lowMemory, (byte) 30);
-            Class49.aPcmStreamMixer_1152 = StaticAudio.method1003(Main.signlink, MouseHandler.gameCanvas);
-            Class55.aDecimator_1289 = new Decimator(22050, StaticAudio.sampleRate);
-            Class40_Sub5_Sub6.loadingPercent = 50;
+        } else if (MovedStatics.loadingPercent == 45) {
+            StaticAudio.method446(Main.signlink, 0, !VertexNormal.lowMemory);
+            StaticAudio.aPcmStreamMixer_1152 = StaticAudio.method1003(Main.signlink, MouseHandler.gameCanvas);
+            StaticAudio.aDecimator_1289 = new Decimator(22050, StaticAudio.sampleRate);
+            MovedStatics.loadingPercent = 50;
             Native.currentLoadingText = English.preparedSoundEngine;
             MovedStatics.anInt1607 = 35;
-        } else if (Class40_Sub5_Sub6.loadingPercent == 50) {
+        } else if (MovedStatics.loadingPercent == 50) {
             int i = 0;
             if (TypeFace.fontSmall != null)
                 i++;
@@ -136,11 +135,11 @@ public class Class40_Sub3 extends Node {
                 Native.currentLoadingText = English.loadingFonts + (i * 100 / 3) + Native.percent;
                 MovedStatics.anInt1607 = 40;
             } else {
-                Class40_Sub5_Sub6.loadingPercent = 60;
+                MovedStatics.loadingPercent = 60;
                 MovedStatics.anInt1607 = 40;
                 Native.currentLoadingText = English.loadedFonts;
             }
-        } else if (Class40_Sub5_Sub6.loadingPercent == 60) {
+        } else if (MovedStatics.loadingPercent == 60) {
             int i = Class60.method988(CacheArchive.huffmanCacheArchive, CacheArchive.gameImageCacheArchive);
             int i_3_ = ISAAC.method288((byte) 119);
             if (i < i_3_) {
@@ -151,9 +150,9 @@ public class Class40_Sub3 extends Node {
                 Native.currentLoadingText = English.loadedTitleScreen;
                 MovedStatics.anInt1607 = 50;
                 MovedStatics.processGameStatus(5);
-                Class40_Sub5_Sub6.loadingPercent = 70;
+                MovedStatics.loadingPercent = 70;
             }
-        } else if (Class40_Sub5_Sub6.loadingPercent == 70) {
+        } else if (MovedStatics.loadingPercent == 70) {
             if (CacheArchive.gameDefinitionsCacheArchive.method185((byte) 98)) {
                 OverlayDefinition.initializeOverlayDefinitionCache(CacheArchive.gameDefinitionsCacheArchive);
                 UnderlayDefinition.initializeUnderlayDefinitionCache(CacheArchive.gameDefinitionsCacheArchive);
@@ -166,14 +165,14 @@ public class Class40_Sub3 extends Node {
                 VarbitDefinition.initializeVarbitDefinitionCache(CacheArchive.gameDefinitionsCacheArchive);
                 VarPlayerDefinition.initializeVarPlayerDefinitionCache(CacheArchive.gameDefinitionsCacheArchive);
                 GameInterface.createInterfaceMemoryBuffers();
-                Class40_Sub5_Sub6.loadingPercent = 80;
+                MovedStatics.loadingPercent = 80;
                 MovedStatics.anInt1607 = 60;
                 Native.currentLoadingText = English.loadedConfig;
             } else {
                 Native.currentLoadingText = English.loadingConfig + CacheArchive.gameDefinitionsCacheArchive.method202() + Native.percent;
                 MovedStatics.anInt1607 = 60;
             }
-        } else if (Class40_Sub5_Sub6.loadingPercent == 80) {
+        } else if (MovedStatics.loadingPercent == 80) {
             int i = 0;
             if (AnimationSequence.minimapCompass != null)
                 i++;
@@ -245,9 +244,9 @@ public class Class40_Sub3 extends Node {
                 MovedStatics.aClass40_Sub5_Sub14_Sub2Array2301[0].mixPalette(i_4_ + i_6_, i_5_ + i_6_, i_6_ + i_7_);
                 Native.currentLoadingText = English.loadedSprites;
                 MovedStatics.anInt1607 = 70;
-                Class40_Sub5_Sub6.loadingPercent = 85;
+                MovedStatics.loadingPercent = 85;
             }
-        } else if (Class40_Sub5_Sub6.loadingPercent == 85) {
+        } else if (MovedStatics.loadingPercent == 85) {
             int i = KeyFocusListener.method955(CacheArchive.gameImageCacheArchive);
             int i_9_ = ActorDefinition.method576();
             if (i < i_9_) {
@@ -255,45 +254,45 @@ public class Class40_Sub3 extends Node {
                 MovedStatics.anInt1607 = 80;
             } else {
                 Native.currentLoadingText = English.loadedGamescreen;
-                Class40_Sub5_Sub6.loadingPercent = 90;
+                MovedStatics.loadingPercent = 90;
                 MovedStatics.anInt1607 = 80;
             }
-        } else if (Class40_Sub5_Sub6.loadingPercent == 90) {
+        } else if (MovedStatics.loadingPercent == 90) {
             if (CacheArchive.gameTextureCacheArchive.method185((byte) 62)) {
                 Class35 class35 = new Class35(CacheArchive.gameTextureCacheArchive, CacheArchive.gameImageCacheArchive, 20, 0.8, !VertexNormal.lowMemory ? 128 : 64);
                 Rasterizer3D.method703(class35);
                 Rasterizer3D.method711(0.8);
-                Class40_Sub5_Sub6.loadingPercent = 110;
+                MovedStatics.loadingPercent = 110;
                 MovedStatics.anInt1607 = 90;
                 Native.currentLoadingText = English.loadedTextures;
             } else {
                 Native.currentLoadingText = English.loadingTextures + CacheArchive.gameTextureCacheArchive.method202() + Native.percent;
                 MovedStatics.anInt1607 = 90;
             }
-        } else if (Class40_Sub5_Sub6.loadingPercent == 110) {
+        } else if (MovedStatics.loadingPercent == 110) {
             Class12.mouseCapturer = new Class39();
             Main.signlink.createThreadNode(10, Class12.mouseCapturer);
             Native.currentLoadingText = English.loadedInputHandler;
-            Class40_Sub5_Sub6.loadingPercent = 120;
+            MovedStatics.loadingPercent = 120;
             MovedStatics.anInt1607 = 94;
-        } else if (Class40_Sub5_Sub6.loadingPercent == 120) {
+        } else if (MovedStatics.loadingPercent == 120) {
             if (CacheArchive.huffmanCacheArchive.method194(Native.huffman, "")) {
                 HuffmanEncoding huffmanEncoding = new HuffmanEncoding(CacheArchive.huffmanCacheArchive.method170("", Native.huffman));
                 Item.method778(huffmanEncoding);
-                Class40_Sub5_Sub6.loadingPercent = 130;
+                MovedStatics.loadingPercent = 130;
                 Native.currentLoadingText = English.loadedWordpack;
                 MovedStatics.anInt1607 = 96;
             } else {
                 Native.currentLoadingText = English.loadingWordPack + Native.aClass1_1348;
                 MovedStatics.anInt1607 = 96;
             }
-        } else if (Class40_Sub5_Sub6.loadingPercent == 130) {
+        } else if (MovedStatics.loadingPercent == 130) {
             if (!CacheArchive.gameInterfaceCacheArchive.method185((byte) 66)) {
                 Native.currentLoadingText = English.loadingInterfaces + (4 * CacheArchive.gameInterfaceCacheArchive.method202() / 5) + Native.percent;
                 MovedStatics.anInt1607 = 100;
             } else if (CacheArchive.clientScriptCacheArchive.method185((byte) 69)) {
                 Native.currentLoadingText = English.loadedInterfaces;
-                Class40_Sub5_Sub6.loadingPercent = 140;
+                MovedStatics.loadingPercent = 140;
                 MovedStatics.anInt1607 = 100;
             } else {
                 Native.currentLoadingText = English.loadingInterfaces + (80 + CacheArchive.clientScriptCacheArchive.method202() / 5) + Native.percent;
@@ -302,7 +301,7 @@ public class Class40_Sub3 extends Node {
         } else {
             if (arg0 != 100)
                 aClass40_Sub5_Sub14_Sub4Array2019 = null;
-            if (Class40_Sub5_Sub6.loadingPercent == 140)
+            if (MovedStatics.loadingPercent == 140)
                 MovedStatics.processGameStatus(10);
         }
     }
