+++ src/main/java/org/runejs/client/Class49.java
@@ -5,12 +5,10 @@ import org.runejs.client.input.MouseHandler;
 import org.runejs.client.media.renderable.Model;
 import org.runejs.client.media.renderable.actor.Player;
 import org.runejs.client.scene.SceneCluster;
-import org.runejs.client.sound.PcmStreamMixer;
 
 public class Class49 {
     public static volatile int anInt1147 = 0;
     public static CacheArchive aCacheArchive_1150;
-    public static PcmStreamMixer aPcmStreamMixer_1152;
     public static int anInt1154;
 
     public static void setCameraPosition(int pitch, int x, int z, int yaw, int y, int ___pitch, int arg6) {
@@ -40,7 +38,7 @@ public class Class49 {
         SceneCluster.cameraZ = -zOffset + z;
         Class12.cameraX = -xOffset + x;
         Class26.anInt627 = pitch;
-        Class40_Sub5_Sub6.cameraY = y + -yOffset;
+        MovedStatics.cameraY = y + -yOffset;
     }
