+++ src/main/java/org/runejs/client/Class57.java
@@ -1,6 +1,5 @@
 package org.runejs.client;
 
-import org.runejs.client.cache.CacheArchive;
 import org.runejs.client.cache.media.IndexedImage;
 
 public class Class57 {
@@ -10,11 +9,4 @@ public class Class57 {
     public static int anInt1342 = 0;
     public static IndexedImage bottomChatBack;
     public static int[] anIntArray1347;
-
-    public static void method975(int songTimeout, int songId) {
-        if(RSCanvas.musicVolume != 0 && songId != -1) {
-            Class33.method414(false, 1, songId, RSCanvas.musicVolume, 0, CacheArchive.jingleCacheArchive);
-            Class35.songTimeout = songTimeout;
-        }
-    }
 }
