+++ src/main/java/org/runejs/client/MovedStatics.java
@@ -5,6 +5,7 @@ import org.runejs.client.cache.media.gameInterface.GameInterface;
 import org.runejs.client.cache.media.gameInterface.GameInterfaceType;
 import org.runejs.client.frame.ScreenController;
 import org.runejs.client.frame.ScreenMode;
+import org.runejs.client.frame.console.Console;
 import org.runejs.client.frame.tab.parts.TabParts;
 import org.runejs.client.input.MouseHandler;
 import org.runejs.client.io.Buffer;
@@ -18,6 +19,7 @@ import org.runejs.client.node.NodeCache;
 import org.runejs.client.cache.CacheArchive;
 import org.runejs.client.cache.media.AnimationSequence;
 import org.runejs.client.cache.media.IndexedImage;
+import org.runejs.client.cache.media.TypeFace;
 import org.runejs.client.frame.ChatBox;
 import org.runejs.client.input.KeyFocusListener;
 import org.runejs.client.language.English;
@@ -33,6 +35,7 @@ import org.runejs.client.scene.InteractiveObject;
 import org.runejs.client.scene.Scene;
 import org.runejs.client.scene.SceneCluster;
 import org.runejs.client.scene.util.CollisionMap;
+import org.runejs.client.sound.StaticAudio;
 import org.runejs.client.util.BitUtils;
 import org.runejs.client.util.Signlink;
 import org.runejs.client.util.SignlinkNode;
@@ -55,7 +58,6 @@ public class MovedStatics {
     public static ProducingGraphicsBuffer tabPieceUpperRight;
     public static boolean aBoolean1444 = false;
     public static int[] anIntArray1445;
-    public static LinkedList aLinkedList_2268 = new LinkedList();
     public static ProducingGraphicsBuffer chatboxRight;
     public static int crossY = 0;
     public static int anInt2280 = 0;
@@ -178,7 +180,7 @@ public class MovedStatics {
             KeyFocusListener.aProducingGraphicsBuffer_1285 = null;
             Class17.aProducingGraphicsBuffer_463 = null;
             anIntArray1013 = null;
-            method405(10);
+            StaticAudio.method405(10);
             GameShell.method19(true);
             ISAAC.aBoolean512 = false;
         }
@@ -189,7 +191,6 @@ public class MovedStatics {
     public static int[] anIntArray889 = new int[128];
     public static int anInt892;
     public static boolean showSidePanelRedrawnText = false;
-    public static int[] anIntArray1916 = new int[50];
     public static IndexedImage aClass40_Sub5_Sub14_Sub2_1919;
     public static int anInt1923 = 0;
 
@@ -230,26 +231,6 @@ public class MovedStatics {
             }
             if (arg0 != 9767)
                 anInt1923 = 48;
-        }
-    }
-
-    public static boolean method446(Signlink arg0, int arg1, boolean arg2, byte arg3) {
-        if (arg3 != 30)
-            Pathfinding.doWalkTo(47, -7, 96, -97, -111, true, -110, 57, -36, -84, -42);
-        if (!Class39.method452(arg0, arg2))
-            return false;
-        if (arg1 > 0)
-            RSString.aClass9_1684 = new NodeCache(arg1);
-        return true;
-    }
-
-    public static void method450(byte arg0) {
-        if (Player.headIconDrawType == 2) {
-            if (arg0 >= -28)
-                method445(-128);
-            Class22_Sub1.method312(2 * ActorDefinition.anInt2404, Class35.anInt1730 + (-Class26.baseY + anInt175 << 7), (ProducingGraphicsBuffer.anInt1637 + -baseX << 7) + Landscape.anInt1170, 4976905);
-            if (ISAAC.anInt522 > -1 && pulseCycle % 20 < 10)
-                aClass40_Sub5_Sub14_Sub4Array2567[0].drawImage(ISAAC.anInt522 + -12, -28 + Class44.anInt1048);
         }
     }
 
@@ -358,8 +339,8 @@ public class MovedStatics {
         }
         if (arg0 != 256)
             anInt2280 = 44;
-        int i = Class37.getFloorDrawHeight(Player.worldLevel, Class12.cameraX, Class40_Sub5_Sub6.cameraY);
-        if (i + -SceneCluster.cameraZ < 800 && (OverlayDefinition.tile_flags[Player.worldLevel][Class12.cameraX >> 7][Class40_Sub5_Sub6.cameraY >> 7] & 0x4) != 0)
+        int i = Class37.getFloorDrawHeight(Player.worldLevel, Class12.cameraX, MovedStatics.cameraY);
+        if (i + -SceneCluster.cameraZ < 800 && (OverlayDefinition.tile_flags[Player.worldLevel][Class12.cameraX >> 7][MovedStatics.cameraY >> 7] & 0x4) != 0)
             return Player.worldLevel;
         return 3;
     }
@@ -617,7 +598,7 @@ public class MovedStatics {
     }
 
     public static void method188() {
-        Class40_Sub5_Sub6.aClass9_2439.clear();
+        MovedStatics.aClass9_2439.clear();
         aClass9_998.clear();
     }
 
@@ -670,14 +651,14 @@ public class MovedStatics {
         for (int i_14_ = 0; i_14_ < i + -1; i_14_++)
             Class17.anIntArray466[i_14_] = Class17.anIntArray466[i_14_ + 1];
         Class17.anIntArray466[i - arg0] = (int) (16.0 * Math.sin((double) pulseCycle / 14.0) + 14.0 * Math.sin((double) pulseCycle / 15.0) + 12.0 * Math.sin((double) pulseCycle / 16.0));
-        if (Class40_Sub5_Sub6.anInt2452 > 0)
-            Class40_Sub5_Sub6.anInt2452 -= 4;
+        if (MovedStatics.anInt2452 > 0)
+            MovedStatics.anInt2452 -= 4;
         if (IdentityKit.anInt2613 > 0)
             IdentityKit.anInt2613 -= 4;
-        if (Class40_Sub5_Sub6.anInt2452 == 0 && IdentityKit.anInt2613 == 0) {
+        if (MovedStatics.anInt2452 == 0 && IdentityKit.anInt2613 == 0) {
             int i_15_ = (int) (2000.0 * Math.random());
             if (i_15_ == 0)
-                Class40_Sub5_Sub6.anInt2452 = 1024;
+                MovedStatics.anInt2452 = 1024;
             if (i_15_ == 1)
                 IdentityKit.anInt2613 = 1024;
         }
@@ -1113,14 +1094,6 @@ public class MovedStatics {
         }
     }
 
-    public static synchronized void method405(int arg1) {
-        if(Class24.method340()) {
-            method557(arg1);
-            Class22_Sub1.musicCacheArchive = null;
-            Class26.aBoolean618 = false;
-        }
-    }
-
     public static void method557(int arg0) {
         method886(0, 0, false, null, arg0);
     }
@@ -1243,10 +1216,10 @@ public class MovedStatics {
 	    for(/**/; i < 4; i++)
 	        Landscape.currentCollisionMap[i].reset();
 	    System.gc();
-	    method405(10);
-	    Class35.songTimeout = 0;
-	    MouseHandler.currentSongId = -1;
-	    Class37.method436();
+	    StaticAudio.method405(10);
+	    StaticAudio.songTimeout = 0;
+	    StaticAudio.currentSongId = -1;
+	    StaticAudio.method436();
 	    processGameStatus(10);
 	}
 
@@ -1546,4 +1519,302 @@ public class MovedStatics {
 	public static int connectionStage = 0;
 	public static int anInt292 = 0;
 	public static int[] anIntArray297 = new int[5];
+
+	public static void drawMenu(int xOffSet, int yOffSet) {
+	    int height = CollisionMap.menuHeight;
+	    int width = VertexNormal.menuWidth;
+	    int offsetX = InteractiveObject.menuOffsetX - (xOffSet);
+	    int offsetY = (-yOffSet) + Main.menuOffsetY;
+	    int colour = 0x5d5447;
+	    ChatBox.redrawChatbox = true;
+	    GameInterface.redrawTabArea = true;
+	    Rasterizer.drawFilledRectangleAlpha(offsetX, offsetY, width, height, colour, 120);
+	    Rasterizer.drawFilledRectangle(1 + offsetX, 1 + offsetY, -2 + width, 16, 0);
+	    Rasterizer.drawUnfilledRectangle(offsetX + 1, 18 + offsetY, -2 + width, -19 + height, 0);
+	    TypeFace.fontBold.drawString(English.chooseOption, offsetX + 3, 14 + offsetY, colour);
+	    int x = Class13.mouseX - (xOffSet);
+	    int y = (-yOffSet) + Landscape.mouseY;
+	    if(Projectile.menuScreenArea == 0) {
+	        x -= 4;
+	        y -= 4;
+	    }
+	    if(Projectile.menuScreenArea == 1) {
+	        y -= 205;
+	        x -= 553;
+	    }
+	    if(Projectile.menuScreenArea == 2) {
+	        y -= 357;
+	        x -= 17;
+	    }
+	    for(int action = 0; action < ActorDefinition.menuActionRow; action++) {
+	        int actionY = 15 * (-action + ActorDefinition.menuActionRow + -1) + offsetY + 31;
+	        int actionColour = 16777215;
+	        if(x > offsetX && offsetX + width > x && y > -13 + actionY && actionY + 3 > y)
+	            actionColour = 16776960;
+	        TypeFace.fontBold.drawShadowedString(Landscape.menuActionTexts[action], offsetX + 3, actionY, true, actionColour);
+	    }
+	}
+
+	public static void manageTextInputs() {
+	    while(method416((byte) -125)) {
+	        if(ItemDefinition.anInt2854 == 28) {
+	            break;
+	        }
+	        if(Console.console.consoleOpen) {
+	            Console.console.handleInput();
+	            break;
+	        }
+	        if(GameInterface.gameScreenInterfaceId != -1 && HuffmanEncoding.reportAbuseInterfaceID == GameInterface.gameScreenInterfaceId) {
+	            if(ItemDefinition.anInt2854 == 85 && Native.reportedName.length() > 0)
+	                Native.reportedName = Native.reportedName.substring(0, -1 + Native.reportedName.length());
+	            if((Class40_Sub5_Sub15.method735(Class59.anInt1388) || Class59.anInt1388 == 32) && Native.reportedName.length() < 12)
+	                Native.reportedName = Native.reportedName + (char) Class59.anInt1388;
+	        } else if(ChatBox.messagePromptRaised) {
+	            if(ItemDefinition.anInt2854 == 85 && ChatBox.chatMessage.length() > 0) {
+	                ChatBox.chatMessage = ChatBox.chatMessage.substring(0, -1 + ChatBox.chatMessage.length());
+	                ChatBox.redrawChatbox = true;
+	            }
+	            if(Player.method793((byte) 120, Class59.anInt1388) && ChatBox.chatMessage.length() < 80) {
+	                ChatBox.chatMessage = ChatBox.chatMessage + (char) Class59.anInt1388;
+	                ChatBox.redrawChatbox = true;
+	            }
+	            if(ItemDefinition.anInt2854 == 84) {
+	                ChatBox.messagePromptRaised = false;
+	                ChatBox.redrawChatbox = true;
+	                if(Class37.anInt876 == 1) {
+	                    long l = RSString.nameToLong(ChatBox.chatMessage);
+	                    addFriend(l);
+	                }
+	                if(Class37.anInt876 == 2 && Player.friendsCount > 0) {
+	                    long l = RSString.nameToLong(ChatBox.chatMessage);
+	                    GameShell.method28(l);
+	                }
+	                if(Class37.anInt876 == 3 && ChatBox.chatMessage.length() > 0) {
+	                    SceneCluster.packetBuffer.putPacket(207);
+	                    SceneCluster.packetBuffer.putByte(0);
+	                    int i = SceneCluster.packetBuffer.currentPosition;
+	                    SceneCluster.packetBuffer.putLongBE(PacketBuffer.aLong2241);
+	                    ChatBox.filterInput();
+	                    ProducingGraphicsBuffer_Sub1.method1052(ChatBox.chatMessage, SceneCluster.packetBuffer);
+	                    SceneCluster.packetBuffer.finishVarByte(-i + SceneCluster.packetBuffer.currentPosition);
+	                    if(ChatBox.privateChatMode == 2) {
+	                        ChatBox.privateChatMode = 1;
+	                        redrawChatbox = true;
+	                        SceneCluster.packetBuffer.putPacket(32);
+	                        SceneCluster.packetBuffer.putByte(ChatBox.publicChatMode);
+	                        SceneCluster.packetBuffer.putByte(ChatBox.privateChatMode);
+	                        SceneCluster.packetBuffer.putByte(ChatBox.tradeMode);
+	                    }
+	                }
+	                if(Class37.anInt876 == 4 && anInt1008 < 100) {
+	                    long l = RSString.nameToLong(ChatBox.chatMessage);
+	                    Class17.method275(l);
+	                }
+	                if(Class37.anInt876 == 5 && anInt1008 > 0) {
+	                    long l = RSString.nameToLong(ChatBox.chatMessage);
+	                    method838(0, l);
+	                }
+	            }
+	        } else if(ChatBox.inputType == 1) {
+	            if(ItemDefinition.anInt2854 == 85 && ChatBox.inputMessage.length() > 0) {
+	                ChatBox.inputMessage = ChatBox.inputMessage.substring(0, ChatBox.inputMessage.length() - 1);
+	                ChatBox.redrawChatbox = true;
+	            }
+	            if(HuffmanEncoding.method1027(Class59.anInt1388) && ChatBox.inputMessage.length() < 10) {
+	                ChatBox.inputMessage = ChatBox.inputMessage + (char) Class59.anInt1388;
+	                ChatBox.redrawChatbox = true;
+	            }
+	            if(ItemDefinition.anInt2854 == 84) {
+	                if(ChatBox.inputMessage.length() > 0) {
+	                    int inputValue = 0;
+	                    try {
+	                        inputValue = Integer.parseInt(ChatBox.inputMessage);
+	                    } catch(Exception _ex) {
+	                        /* empty */
+	                    }
+	                    SceneCluster.packetBuffer.putPacket(238);
+	                    SceneCluster.packetBuffer.putIntBE(inputValue);
+	                }
+	                ChatBox.redrawChatbox = true;
+	                ChatBox.inputType = 0;
+	            }
+	        } else if(ChatBox.inputType == 2) {
+	            if(ItemDefinition.anInt2854 == 85 && ChatBox.inputMessage.length() > 0) {
+	                ChatBox.inputMessage = ChatBox.inputMessage.substring(0, -1 + ChatBox.inputMessage.length());
+	                ChatBox.redrawChatbox = true;
+	            }
+	            if((Class40_Sub5_Sub15.method735(Class59.anInt1388) || Class59.anInt1388 == 32) && ChatBox.inputMessage.length() < 12) {
+	                ChatBox.inputMessage = ChatBox.inputMessage + (char) Class59.anInt1388;
+	                ChatBox.redrawChatbox = true;
+	            }
+	            if(ItemDefinition.anInt2854 == 84) {
+	                if(ChatBox.inputMessage.length() > 0) {
+	                    SceneCluster.packetBuffer.putPacket(86);
+	                    SceneCluster.packetBuffer.putLongBE(RSString.nameToLong(ChatBox.inputMessage));
+	                }
+	                ChatBox.inputType = 0;
+	                ChatBox.redrawChatbox = true;
+	            }
+	        } else if(ChatBox.inputType == 3) {
+	            if(ItemDefinition.anInt2854 == 85 && ChatBox.inputMessage.length() > 0) {
+	                ChatBox.inputMessage = ChatBox.inputMessage.substring(0, ChatBox.inputMessage.length() - 10);
+	                ChatBox.redrawChatbox = true;
+	            }
+	            if(Player.method793((byte) 122, Class59.anInt1388) && ChatBox.inputMessage.length() < 40) {
+	                ChatBox.inputMessage = ChatBox.inputMessage + (char) Class59.anInt1388;
+	                ChatBox.redrawChatbox = true;
+	            }
+	        } else if(GameInterface.chatboxInterfaceId == -1 && GameInterface.fullscreenInterfaceId == -1) {
+	            if(ItemDefinition.anInt2854 == 85 && ChatBox.chatboxInput.length() > 0) {
+	                ChatBox.chatboxInput = ChatBox.chatboxInput.substring(0, ChatBox.chatboxInput.length() - 1);
+	                ChatBox.redrawChatbox = true;
+	            }
+	            if(Player.method793((byte) 124, Class59.anInt1388) && ChatBox.chatboxInput.length() < 80) {
+	                ChatBox.chatboxInput = ChatBox.chatboxInput + (char) Class59.anInt1388;
+	                ChatBox.redrawChatbox = true;
+	            }
+	            if(ItemDefinition.anInt2854 == 84 && ChatBox.chatboxInput.length() > 0) {
+	                if(InteractiveObject.playerRights > 1) {
+	                    if(ChatBox.chatboxInput.equals(English.commandClientDrop))
+	                        Class59.dropClient();
+	                    if(ChatBox.chatboxInput.equals(English.commandFpson)) {
+	                        InteractiveObject.showFps = true;
+	                        ChatBox.inputType = 3;
+	                    }
+	                    if(ChatBox.chatboxInput.startsWith("::region")) {
+	                        for(int qq = 0; qq < 469; qq++) {
+	                            if(GameInterface.decodeGameInterface(qq)) {
+	                                GameInterface[] gameInterfaces = GameInterface.cachedInterfaces[qq];
+	                                for(int y = 0; gameInterfaces.length > y; y++) {
+	                                    GameInterface gameInterface = gameInterfaces[y];
+	                                    if(gameInterface.disabledText != null) {
+	                                        String text = gameInterface.disabledText.toString().toLowerCase();
+	                                        if(gameInterface.disabledText.toString().toLowerCase().contains("bank")) {
+	                                            System.out.println(qq + " contains " + text);
+	                                        }
+	                                    }
+	                                }
+	                            }
+	                        }
+	                    }
+	
+	                    if(ChatBox.chatboxInput.equals(Native.cmd_fpsoff))
+	                        InteractiveObject.showFps = false;
+	                    if(ChatBox.chatboxInput.equals(English.commandNoclip)) {
+	                        for(int i = 0; i < 4; i++) {
+	                            for(int i_9_ = 1; i_9_ < 103; i_9_++) {
+	                                for(int i_10_ = 1; i_10_ < 103; i_10_++)
+	                                    Landscape.currentCollisionMap[i].clippingData[i_9_][i_10_] = 0;
+	                            }
+	                        }
+	                    }
+	                    if(ChatBox.chatboxInput.equals(English.commandErrorTest) && Class44.modewhere == 2)
+	                        throw new RuntimeException();
+	                    if(ChatBox.chatboxInput.equals(Native.cmd_hiddenbuttontest))
+	                        PacketBuffer.hiddenButtonTest = true;
+	                }
+	                if(ChatBox.chatboxInput.startsWith(Native.cmd_prefix)) {
+	                    SceneCluster.packetBuffer.putPacket(248);
+	                    SceneCluster.packetBuffer.putByte(-1 + ChatBox.chatboxInput.length());
+	                    SceneCluster.packetBuffer.putString(ChatBox.chatboxInput.substring(2));
+	                } else {
+	                    int i = 0;
+	                    String class1 = ChatBox.chatboxInput.toLowerCase();
+	                    int i_11_ = 0;
+	                    if(class1.startsWith(English.effectYellow)) {
+	                        i_11_ = 0;
+	                        ChatBox.chatboxInput = ChatBox.chatboxInput.substring(English.effectYellow.length());
+	                    } else if(class1.startsWith(English.effectRed)) {
+	                        ChatBox.chatboxInput = ChatBox.chatboxInput.substring(English.effectRed.length());
+	                        i_11_ = 1;
+	                    } else if(class1.startsWith(English.effectGreen)) {
+	                        i_11_ = 2;
+	                        ChatBox.chatboxInput = ChatBox.chatboxInput.substring(English.effectGreen.length());
+	                    } else if(class1.startsWith(English.effectCyan)) {
+	                        i_11_ = 3;
+	                        ChatBox.chatboxInput = ChatBox.chatboxInput.substring(English.effectCyan.length());
+	                    } else if(class1.startsWith(English.effectPurple)) {
+	                        i_11_ = 4;
+	                        ChatBox.chatboxInput = ChatBox.chatboxInput.substring(English.effectPurple.length());
+	                    } else if(class1.startsWith(English.effectWhite)) {
+	                        ChatBox.chatboxInput = ChatBox.chatboxInput.substring(English.effectWhite.length());
+	                        i_11_ = 5;
+	                    } else if(class1.startsWith(Native.aClass1_1019)) {
+	                        i_11_ = 6;
+	                        ChatBox.chatboxInput = ChatBox.chatboxInput.substring(Native.aClass1_1019.length());
+	                    } else if(class1.startsWith(English.effectFlash2)) {
+	                        i_11_ = 7;
+	                        ChatBox.chatboxInput = ChatBox.chatboxInput.substring(English.effectFlash2.length());
+	                    } else if(class1.startsWith(English.effectFlash3)) {
+	                        i_11_ = 8;
+	                        ChatBox.chatboxInput = ChatBox.chatboxInput.substring(English.effectFlash3.length());
+	                    } else if(class1.startsWith(English.effectGlow1)) {
+	                        i_11_ = 9;
+	                        ChatBox.chatboxInput = ChatBox.chatboxInput.substring(English.effectGlow1.length());
+	                    } else if(class1.startsWith(English.effectGlow2)) {
+	                        ChatBox.chatboxInput = ChatBox.chatboxInput.substring(English.effectGlow2.length());
+	                        i_11_ = 10;
+	                    } else if(class1.startsWith(English.effectGlow3)) {
+	                        ChatBox.chatboxInput = ChatBox.chatboxInput.substring(English.effectGlow3.length());
+	                        i_11_ = 11;
+	                    }
+	                    class1 = ChatBox.chatboxInput.toLowerCase();
+	                    if(class1.startsWith(English.effectWave)) {
+	                        i = 1;
+	                        ChatBox.chatboxInput = ChatBox.chatboxInput.substring(English.effectWave.length());
+	                    } else if(class1.startsWith(English.effectWave2)) {
+	                        i = 2;
+	                        ChatBox.chatboxInput = ChatBox.chatboxInput.substring(English.effectWave2.length());
+	                    } else if(class1.startsWith(English.effectShake)) {
+	                        ChatBox.chatboxInput = ChatBox.chatboxInput.substring(English.effectShake.length());
+	                        i = 3;
+	                    } else if(class1.startsWith(Native.aClass1_2438)) {
+	                        i = 4;
+	                        ChatBox.chatboxInput = ChatBox.chatboxInput.substring(Native.aClass1_2438.length());
+	                    } else if(class1.startsWith(English.effectSlide)) {
+	                        i = 5;
+	                        ChatBox.chatboxInput = ChatBox.chatboxInput.substring(English.effectSlide.length());
+	                    }
+	                    SceneCluster.packetBuffer.putPacket(75);
+	                    SceneCluster.packetBuffer.putByte(0);
+	                    int i_12_ = SceneCluster.packetBuffer.currentPosition;
+	                    SceneCluster.packetBuffer.putByte(i_11_);
+	                    SceneCluster.packetBuffer.putByte(i);
+	                    ChatBox.filterInput();
+	                    ProducingGraphicsBuffer_Sub1.method1052(ChatBox.chatboxInput, SceneCluster.packetBuffer);
+	                    SceneCluster.packetBuffer.finishVarByte(SceneCluster.packetBuffer.currentPosition + -i_12_);
+	                    if(ChatBox.publicChatMode == 2) {
+	                        redrawChatbox = true;
+	                        ChatBox.publicChatMode = 3;
+	                        SceneCluster.packetBuffer.putPacket(32);
+	                        SceneCluster.packetBuffer.putByte(ChatBox.publicChatMode);
+	                        SceneCluster.packetBuffer.putByte(ChatBox.privateChatMode);
+	                        SceneCluster.packetBuffer.putByte(ChatBox.tradeMode);
+	                    }
+	                }
+	                ChatBox.redrawChatbox = true;
+	                ChatBox.chatboxInput = "";
+	            }
+	        }
+	    }
+	
+	}
+
+	private static char VALID_CHARACTERS[] = {' ', 'e', 't', 'a', 'o', 'i', 'h', 'n', 's', 'r', 'd', 'l', 'u', 'm', 'w',
+	'c', 'y', 'f', 'g', 'p', 'b', 'v', 'k', 'x', 'j', 'q', 'z', '0', '1', '2', '3', '4', '5', '6', '7', '8',
+	'9', ' ', '!', '?', '.', ',', ':', ';', '(', ')', '-', '&', '*', '\\', '\'', '@', '#', '+', '=', '\243',
+	'$', '%', '"', '[', ']'};
+
+	public static void method586(int arg0) {
+	    ISAAC.cachedActorDefinitions.clear();
+	    aClass9_1611.clear();
+	    if(arg0 <= 63)
+	        English.errorConnectingToServer = null;
+	}
+
+	public static NodeCache aClass9_2439 = new NodeCache(64);
+	public static int currentCameraPositionH;
+	public static int anInt2452 = 0;
+	public static int loadingPercent = 0;
+	public static int cameraY;
 }
