+++ src/main/java/org/runejs/client/cache/def/ItemDefinition.java
@@ -23,7 +23,6 @@ import java.awt.*;
 
 public class ItemDefinition extends CachedNode implements EntityDefinition {
     public static int anInt2798 = 0;
-    public static int[] soundVolume = new int[50];
     public static int selectedMask;
     public static int anInt2846;
     public static int anInt2850 = -1;
