+++ src/main/java/org/runejs/client/frame/ChatBox.java
@@ -180,7 +180,7 @@ public class ChatBox {
                 redrawChatbox = true;
         }
         if(MovedStatics.menuOpen && ScreenController.frameMode == ScreenMode.FIXED && Projectile.menuScreenArea == 2)
-            Class40_Sub5_Sub6.drawMenu(0, 0);
+            MovedStatics.drawMenu(0, 0);
         if(ScreenController.frameMode == ScreenMode.FIXED) {
             LinkedList.drawChatBoxGraphics();
         }
