+++ src/main/java/org/runejs/client/input/MouseHandler.java
@@ -28,7 +28,6 @@ import java.awt.event.*;
 public class MouseHandler implements MouseListener, MouseMotionListener, FocusListener, MouseWheelListener {
     public static int anInt1450 = -1;
     public static NodeCache modelCache = new NodeCache(50);
-    public static int currentSongId = -1;
     public static ImageRGB[] minimapHint = new ImageRGB[1000];
     public static int currentTickSample;
     public static Canvas gameCanvas;
