+++ src/main/java/org/runejs/client/media/renderable/Item.java
@@ -16,6 +16,7 @@ import org.runejs.client.net.ISAAC;
 import org.runejs.client.scene.GroundItemTile;
 import org.runejs.client.scene.tile.SceneTile;
 import org.runejs.client.scene.tile.Wall;
+import org.runejs.client.sound.StaticAudio;
 import org.runejs.client.*;
 import org.runejs.Configuration;
 import org.runejs.whitelabel.Images;
@@ -32,12 +33,12 @@ public class Item extends Renderable {
     public static void calculateCameraPosition() {
         int sceneX = Buffer.cameraOffsetX + Player.localPlayer.worldX;
         int sceneY = Player.localPlayer.worldY + MovedStatics.cameraOffsetY;
-        if (Class40_Sub5_Sub6.currentCameraPositionH - sceneX < -500 || -sceneX + Class40_Sub5_Sub6.currentCameraPositionH > 500 || MovedStatics.currentCameraPositionV + -sceneY < -500 || -sceneY + MovedStatics.currentCameraPositionV > 500) {
+        if (MovedStatics.currentCameraPositionH - sceneX < -500 || -sceneX + MovedStatics.currentCameraPositionH > 500 || MovedStatics.currentCameraPositionV + -sceneY < -500 || -sceneY + MovedStatics.currentCameraPositionV > 500) {
             MovedStatics.currentCameraPositionV = sceneY;
-            Class40_Sub5_Sub6.currentCameraPositionH = sceneX;
+            MovedStatics.currentCameraPositionH = sceneX;
         }
-        if (Class40_Sub5_Sub6.currentCameraPositionH != sceneX)
-            Class40_Sub5_Sub6.currentCameraPositionH += (-Class40_Sub5_Sub6.currentCameraPositionH + sceneX) / 16;
+        if (MovedStatics.currentCameraPositionH != sceneX)
+            MovedStatics.currentCameraPositionH += (-MovedStatics.currentCameraPositionH + sceneX) / 16;
         if (MovedStatics.currentCameraPositionV != sceneY)
             MovedStatics.currentCameraPositionV += (-MovedStatics.currentCameraPositionV + sceneY) / 16;
 
@@ -55,14 +56,14 @@ public class Item extends Renderable {
             Class60.cameraVelocityVertical /= 2;
         int i_1_ = MovedStatics.currentCameraPositionV >> 7;
         GroundItemTile.cameraHorizontal = Wall.cameraVelocityHorizontal / 2 + GroundItemTile.cameraHorizontal & 0x7ff;
-        int i_2_ = Class40_Sub5_Sub6.currentCameraPositionH >> 7;
+        int i_2_ = MovedStatics.currentCameraPositionH >> 7;
         Class65.cameraVertical += Class60.cameraVelocityVertical / 2;
         int i_3_ = 0;
         if (Class65.cameraVertical < 128)
             Class65.cameraVertical = 128;
         if (Class65.cameraVertical > 383)
             Class65.cameraVertical = 383;
-        int i_4_ = Class37.getFloorDrawHeight(Player.worldLevel, Class40_Sub5_Sub6.currentCameraPositionH, MovedStatics.currentCameraPositionV);
+        int i_4_ = Class37.getFloorDrawHeight(Player.worldLevel, MovedStatics.currentCameraPositionH, MovedStatics.currentCameraPositionV);
         if (i_2_ > 3 && i_1_ > 3 && i_2_ < 100 && i_1_ < 100) {
             for (int i_5_ = -4 + i_2_; i_5_ <= 4 + i_2_; i_5_++) {
                 for (int i_6_ = -4 + i_1_; 4 + i_1_ >= i_6_; i_6_++) {
@@ -218,9 +219,9 @@ public class Item extends Renderable {
             MovedStatics.anIntArray178 = new int[32768];
             Class26.loginScreenState = 0;
             if (RSCanvas.musicVolume != 0 && !VertexNormal.lowMemory)
-                Class33.method412(false, CacheArchive.musicCacheArchive, 0, (byte) 66, Native.titleSong, 10, "", RSCanvas.musicVolume);
+                StaticAudio.method412(false, CacheArchive.musicCacheArchive, 0, Native.titleSong, 10, "", RSCanvas.musicVolume);
             else
-                MovedStatics.method405(10);
+                StaticAudio.method405(10);
             GameShell.method19(false);
             MovedStatics.clearScreen = true;
             ISAAC.aBoolean512 = true;
