+++ src/main/java/org/runejs/client/media/renderable/actor/Actor.java
@@ -8,9 +8,9 @@ import org.runejs.client.io.Buffer;
 import org.runejs.client.language.English;
 import org.runejs.client.media.VertexNormal;
 import org.runejs.client.media.renderable.Renderable;
-import org.runejs.client.net.PacketBuffer;
 import org.runejs.client.scene.tile.Wall;
 import org.runejs.client.scene.tile.WallDecoration;
+import org.runejs.client.sound.StaticAudio;
 import org.runejs.client.util.Signlink;
 import org.runejs.client.*;
 
@@ -347,7 +347,7 @@ public abstract class Actor extends Renderable {
                 Class55.destinationY -= i_35_;
             }
             Player.cutsceneActive = false;
-            PacketBuffer.currentSound = 0;
+            StaticAudio.currentSound = 0;
             Class57.aLinkedList_1332.method906(0);
             Class43.projectileQueue.method906(0);
         }
