+++ src/main/java/org/runejs/client/sound/Class40_Sub5_Sub6.java
@@ -1,0 +1,10 @@
+package org.runejs.client.sound;
+
+import org.runejs.client.node.CachedNode;
+
+public class Class40_Sub5_Sub6 extends CachedNode {
+    public byte[] aByteArray2441;
+    public Class40_Sub5_Sub6(byte[] arg0) {
+        aByteArray2441 = arg0;
+    }
+}
